pyproject.toml
[tool.poetry]
name = "tea"
version = "0.1.0"
description = "tea is a tool for monitoring Databases instances, VPN's and Endpoints."
authors = ["leonardo.freua <leonardo.batista.freua@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Leonardofreua/tea"
keywords = ["tea", "monitoring", "metrics"]

[tool.poetry.dependencies]
python = "^3.7"
pyroute2 = "^0.5.7"
psycopg2 = "^2.8.4"
PyYAML = "^5.3"
requests = "^2.22.0"
dynaconf = "^2.2.2"
Flask = "^1.1.1"
Flask-SQLAlchemy = "^2.4.1"
Flask-Admin = "^1.5.4"
Flask-Bootstrap = "^3.3.7"
flask-shell-ipython = "^0.4.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
bandit = "^1.6.2"
isort = "^4.3.21"
flake8 = "^3.7.9"
flake8-bugbear = "^20.1.2"
flask_debugtoolbar = "^0.10.1"
coverage = {extras = ["toml"], version = "^5.0.3"}
pytest = "^5.3.4"
pytest-cov = "^2.8.1"

[tool.poetry.scripts]
tea = "tea.app:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["tea"]

[tool.coverage.report]
show_missing = true
#fail_under = 100

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
